+++ teacher.rb
@@ -3,6 +3,7 @@ require './person'
 class Teacher < Person
   attr_reader :name, :age, :parent_permission
   attr_accessor :specialization, :id
+
   def initialize(age:, specialization:, name: 'Unknown', parent_permission: true)
     super(age: age, parent_permission: parent_permission, name: name)
     @specialization = specialization
